restructure files into components and pages folders, and delete css files
