+++ wwwroot/src/App.js
@@ -9,8 +9,8 @@ import blue from '@material-ui/core/colors/blue';
 import grey from '@material-ui/core/colors/grey';
 
 import './App.css';
-import SignIn from './components/SignIn';
-import SignUp from './components/SignUp';
+import SignIn from './pages/SignIn';
+import SignUp from './pages/SignUp';
 
 const theme = createMuiTheme({
   palette: {
